Move api utils into api
