+++ enterprise/frontend/src/metabase-enterprise/advanced_permissions/services.ts
@@ -1,4 +1,4 @@
-import { GET } from "metabase/utils/api";
+import { GET } from "metabase/api/legacy-client";
 
 export const ImpersonationApi = {
   get: GET("/api/ee/advanced-permissions/impersonation"),
