+++ frontend/src/metabase/api/ai-streaming/requests.unit.spec.ts
@@ -1,6 +1,6 @@
 import fetchMock from "fetch-mock";
 
-import api from "metabase/utils/api";
+import api from "metabase/api/legacy-client";
 
 import {
   aiStreamingQuery,
