+++ frontend/src/metabase/setup/actions.ts
@@ -2,6 +2,7 @@ import { createAction } from "@reduxjs/toolkit";
 import { t } from "ttag";
 
 import { userApi } from "metabase/api";
+import { loadLocalization } from "metabase/api/load-localization";
 import { createDatabase } from "metabase/redux/databases";
 import {
   initializeSettings,
@@ -12,7 +13,6 @@ import type { InviteInfo, Locale, State, UserInfo } from "metabase/redux/store";
 import { createAsyncThunk } from "metabase/redux/utils";
 import { getSetting } from "metabase/selectors/settings";
 import { SetupApi } from "metabase/services";
-import { loadLocalization } from "metabase/utils/i18n";
 import MetabaseSettings from "metabase/utils/settings";
 import type { DatabaseData, Settings, UsageReason } from "metabase-types/api";
