+++ frontend/src/metabase/urls/utils.unit.spec.ts
@@ -1,4 +1,4 @@
-import api from "metabase/utils/api";
+import api from "metabase/api/legacy-client";
 
 import {
   getSubpathSafeUrl,
