+++ frontend/src/metabase/visualizations/components/LeafletTilePinMap.tsx
@@ -1,9 +1,9 @@
 import L from "leaflet";
 import type { ContextType } from "react";
 
+import { GET } from "metabase/api/legacy-client";
 import { EmbeddingEntityContext } from "metabase/embedding/context";
 import { isEmbeddingSdk } from "metabase/embedding-sdk/config";
-import { GET } from "metabase/utils/api";
 import { isWithinIframe } from "metabase/utils/iframe";
 import type { DashboardId } from "metabase-types/api";
